+++ fsw/src/bpnode_adu_in.c
@@ -39,7 +39,7 @@
 int32 BPNode_AduInCreateTasks(void)
 {
     int32  Status;
-    uint8  i;
+    uint32 i;
     char   NameBuff[OS_MAX_API_NAME];
     uint16 TaskPriority;
 
@@ -126,11 +126,11 @@ int32 BPNode_AduInCreateTasks(void)
 }
 
 /* Task initialization for ADU In task(s) */
-int32 BPNode_AduIn_TaskInit(uint8 *ChanId)
+int32 BPNode_AduIn_TaskInit(uint32 *ChanId)
 {
     CFE_ES_TaskId_t TaskId;
     int32           Status;
-    uint8           i;
+    uint32          i;
     char            NameBuff[OS_MAX_API_NAME];
 
     /* Get the task ID of currently running child task */
@@ -203,7 +203,7 @@ void BPNode_AduIn_AppMain(void)
 {
     int32 Status;
     CFE_SB_Buffer_t *BufPtr = NULL;
-    uint8 ChanId = BPLIB_MAX_NUM_CHANNELS; /* Set to garbage value */
+    uint32 ChanId = BPLIB_MAX_NUM_CHANNELS; /* Set to garbage value */
     BPLib_NC_ApplicationState_t AppState;
 
     /* Perform task-specific initialization */
@@ -311,7 +311,7 @@ void BPNode_AduIn_AppMain(void)
 }
 
 /* Exit child task */
-void BPNode_AduIn_TaskExit(uint8 ChanId)
+void BPNode_AduIn_TaskExit(uint32 ChanId)
 {
     BPLib_EM_SendEvent(BPNODE_ADU_IN_EXIT_CRIT_EID, BPLib_EM_EventType_CRITICAL,
                       "[ADU In #%d]: Terminating Task. RunStatus = %d.",
