+++ unit-test/bpnode_adu_out_tests.c
@@ -129,8 +129,8 @@ void Test_BPNode_AduOutCreateTasks_TakeSemErr(void)
 /* Test BPNode_AduOut_TaskInit when everything succeeds */
 void Test_BPNode_AduOut_TaskInit_Nominal(void)
 {
-    uint8 ChanId = BPLIB_MAX_NUM_CHANNELS;
-    uint8 ExpChanId = 0;
+    uint32 ChanId = BPLIB_MAX_NUM_CHANNELS;
+    uint32 ExpChanId = 0;
     CFE_ES_TaskId_t TaskId = 1234;
 
     /* Test setup */
@@ -154,8 +154,8 @@ void Test_BPNode_AduOut_TaskInit_Nominal(void)
 /* Test BPNode_AduOut_TaskInit when it fails to get the task ID */
 void Test_BPNode_AduOut_TaskInit_GetIdErr(void)
 {
-    uint8 ChanId = BPLIB_MAX_NUM_CHANNELS;
-    uint8 ExpChanId = BPLIB_MAX_NUM_CHANNELS;
+    uint32 ChanId = BPLIB_MAX_NUM_CHANNELS;
+    uint32 ExpChanId = BPLIB_MAX_NUM_CHANNELS;
 
     /* Test setup */
     UT_SetDeferredRetcode(UT_KEY(CFE_ES_GetTaskID), 1, CFE_ES_ERR_RESOURCEID_NOT_VALID);
@@ -174,8 +174,8 @@ void Test_BPNode_AduOut_TaskInit_GetIdErr(void)
 /* Test BPNode_AduOut_TaskInit when the task IDs don't match */
 void Test_BPNode_AduOut_TaskInit_MatchIdErr(void)
 {
-    uint8 ChanId = BPLIB_MAX_NUM_CHANNELS;
-    uint8 ExpChanId = BPLIB_MAX_NUM_CHANNELS;
+    uint32 ChanId = BPLIB_MAX_NUM_CHANNELS;
+    uint32 ExpChanId = BPLIB_MAX_NUM_CHANNELS;
     CFE_ES_TaskId_t TaskId = 1234;
 
     /* Test setup */
@@ -197,8 +197,8 @@ void Test_BPNode_AduOut_TaskInit_MatchIdErr(void)
 /* Test BPNode_AduOut_TaskInit when init semaphore give fails */
 void Test_BPNode_AduOut_TaskInit_GiveSemErr(void)
 {
-    uint8 ChanId = BPLIB_MAX_NUM_CHANNELS;
-    uint8 ExpChanId = 0;
+    uint32 ChanId = BPLIB_MAX_NUM_CHANNELS;
+    uint32 ExpChanId = 0;
     CFE_ES_TaskId_t TaskId = 1234;
 
     /* Test setup */
@@ -221,7 +221,7 @@ void Test_BPNode_AduOut_TaskInit_GiveSemErr(void)
 /* Test BPNode_AduOut_AppMain when app state is started */
 void Test_BPNode_AduOut_AppMain_Nominal(void)
 {
-    uint8 ChanId = 0;
+    uint32 ChanId = 0;
     CFE_ES_TaskId_t TaskId = 1234;
 
     /* Test setup */
@@ -244,7 +244,7 @@ void Test_BPNode_AduOut_AppMain_Nominal(void)
 void Test_BPNode_AduOut_AppMain_TakeSemErr(void)
 {
     CFE_ES_TaskId_t TaskId;
-    uint8 ChanId;
+    uint32 ChanId;
 
     /* Force a failed task wakeup */
     UT_SetDeferredRetcode(UT_KEY(OS_BinSemTimedWait), 1, OS_SEM_FAILURE);
@@ -274,7 +274,7 @@ void Test_BPNode_AduOut_AppMain_TakeSemErr(void)
 void Test_BPNode_AduOut_AppMain_WakeupSemTimeout(void)
 {
     CFE_ES_TaskId_t TaskId;
-    uint8 ChanId;
+    uint32 ChanId;
 
     /* Force a failed task wakeup */
     UT_SetDeferredRetcode(UT_KEY(OS_BinSemTimedWait), 1, OS_SEM_TIMEOUT);
@@ -304,7 +304,7 @@ void Test_BPNode_AduOut_AppMain_WakeupSemTimeout(void)
 /* Test BPNode_AduOut_AppMain when initialization failed but channel ID is known */
 void Test_BPNode_AduOut_AppMain_InitErr(void)
 {
-    uint8 ChanId = 0;
+    uint32 ChanId = 0;
     CFE_ES_TaskId_t TaskId = 1234;
 
     /* Test setup */
@@ -341,7 +341,7 @@ void Test_BPNode_AduOut_AppMain_ChanIdErr(void)
 /* Test BPNode_AduOut_AppMain when app state is stopped */
 void Test_BPNode_AduOut_AppMain_AppStopped(void)
 {
-    uint8 ChanId = 0;
+    uint32 ChanId = 0;
     CFE_ES_TaskId_t TaskId = 1234;
 
     /* Test setup */
@@ -362,7 +362,7 @@ void Test_BPNode_AduOut_AppMain_AppStopped(void)
 /* Test BPNode_AduOut_TaskExit in nominal shutdown */
 void Test_BPNode_AduOut_TaskExit_Nominal(void)
 {
-    uint8 ChanId = 0;
+    uint32 ChanId = 0;
 
     UtAssert_VOIDCALL(BPNode_AduOut_TaskExit(ChanId));
