+++ unit-test/stubs/fwp_adup_stubs.c
@@ -32,11 +32,11 @@
  * Generated stub function for BPA_ADUP_AddApplication()
  * ----------------------------------------------------
  */
-BPLib_Status_t BPA_ADUP_AddApplication(uint8_t ChanId)
+BPLib_Status_t BPA_ADUP_AddApplication(uint32_t ChanId)
 {
     UT_GenStub_SetupReturnBuffer(BPA_ADUP_AddApplication, BPLib_Status_t);
 
-    UT_GenStub_AddParam(BPA_ADUP_AddApplication, uint8_t, ChanId);
+    UT_GenStub_AddParam(BPA_ADUP_AddApplication, uint32_t, ChanId);
 
     UT_GenStub_Execute(BPA_ADUP_AddApplication, Basic, NULL);
 
@@ -48,12 +48,12 @@ BPLib_Status_t BPA_ADUP_AddApplication(uint8_t ChanId)
  * Generated stub function for BPA_ADUP_In()
  * ----------------------------------------------------
  */
-BPLib_Status_t BPA_ADUP_In(void *AduPtr, uint8_t ChanId)
+BPLib_Status_t BPA_ADUP_In(void *AduPtr, uint32_t ChanId)
 {
     UT_GenStub_SetupReturnBuffer(BPA_ADUP_In, BPLib_Status_t);
 
     UT_GenStub_AddParam(BPA_ADUP_In, void *, AduPtr);
-    UT_GenStub_AddParam(BPA_ADUP_In, uint8_t, ChanId);
+    UT_GenStub_AddParam(BPA_ADUP_In, uint32_t, ChanId);
 
     UT_GenStub_Execute(BPA_ADUP_In, Basic, NULL);
 
@@ -65,11 +65,11 @@ BPLib_Status_t BPA_ADUP_In(void *AduPtr, uint8_t ChanId)
  * Generated stub function for BPA_ADUP_Out()
  * ----------------------------------------------------
  */
-BPLib_Status_t BPA_ADUP_Out(uint8_t ChanId, uint32_t Timeout)
+BPLib_Status_t BPA_ADUP_Out(uint32_t ChanId, uint32_t Timeout)
 {
     UT_GenStub_SetupReturnBuffer(BPA_ADUP_Out, BPLib_Status_t);
 
-    UT_GenStub_AddParam(BPA_ADUP_Out, uint8_t, ChanId);
+    UT_GenStub_AddParam(BPA_ADUP_Out, uint32_t, ChanId);
     UT_GenStub_AddParam(BPA_ADUP_Out, uint32_t, Timeout);
 
     UT_GenStub_Execute(BPA_ADUP_Out, Basic, NULL);
@@ -82,11 +82,11 @@ BPLib_Status_t BPA_ADUP_Out(uint8_t ChanId, uint32_t Timeout)
  * Generated stub function for BPA_ADUP_RemoveApplication()
  * ----------------------------------------------------
  */
-BPLib_Status_t BPA_ADUP_RemoveApplication(uint8_t ChanId)
+BPLib_Status_t BPA_ADUP_RemoveApplication(uint32_t ChanId)
 {
     UT_GenStub_SetupReturnBuffer(BPA_ADUP_RemoveApplication, BPLib_Status_t);
 
-    UT_GenStub_AddParam(BPA_ADUP_RemoveApplication, uint8_t, ChanId);
+    UT_GenStub_AddParam(BPA_ADUP_RemoveApplication, uint32_t, ChanId);
 
     UT_GenStub_Execute(BPA_ADUP_RemoveApplication, Basic, NULL);
 
@@ -98,11 +98,11 @@ BPLib_Status_t BPA_ADUP_RemoveApplication(uint8_t ChanId)
  * Generated stub function for BPA_ADUP_StartApplication()
  * ----------------------------------------------------
  */
-BPLib_Status_t BPA_ADUP_StartApplication(uint8_t ChanId)
+BPLib_Status_t BPA_ADUP_StartApplication(uint32_t ChanId)
 {
     UT_GenStub_SetupReturnBuffer(BPA_ADUP_StartApplication, BPLib_Status_t);
 
-    UT_GenStub_AddParam(BPA_ADUP_StartApplication, uint8_t, ChanId);
+    UT_GenStub_AddParam(BPA_ADUP_StartApplication, uint32_t, ChanId);
 
     UT_GenStub_Execute(BPA_ADUP_StartApplication, Basic, NULL);
 
@@ -114,11 +114,11 @@ BPLib_Status_t BPA_ADUP_StartApplication(uint8_t ChanId)
  * Generated stub function for BPA_ADUP_StopApplication()
  * ----------------------------------------------------
  */
-BPLib_Status_t BPA_ADUP_StopApplication(uint8_t ChanId)
+BPLib_Status_t BPA_ADUP_StopApplication(uint32_t ChanId)
 {
     UT_GenStub_SetupReturnBuffer(BPA_ADUP_StopApplication, BPLib_Status_t);
 
-    UT_GenStub_AddParam(BPA_ADUP_StopApplication, uint8_t, ChanId);
+    UT_GenStub_AddParam(BPA_ADUP_StopApplication, uint32_t, ChanId);
 
     UT_GenStub_Execute(BPA_ADUP_StopApplication, Basic, NULL);
